Create shared func
